requirements.txt
attrs==24.3.0
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==44.0.0
EditorConfig==0.17.0
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httplib2==0.22.0
hyperframe==6.0.1
idna==3.10
jsbeautifier==1.15.1
kaitaistruct==0.10
loguru==0.7.3
numpy==2.2.1
oauthlib==3.2.2
opencv-python==4.10.0.84
outcome==1.3.0.post0
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.29.2
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydivert==2.1.0
pyOpenSSL==24.3.0
pyparsing==3.2.0
PySocks==1.7.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
selenium==4.27.1
selenium-wire==5.1.0
blinker==1.7.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
websocket-client==1.8.0
win32_setctime==1.2.0
wsproto==1.2.0
zstandard==0.23.0